fixed killUpdate which was removing old library

+++ spotify-mini-player/src/functions.php
@@ -5957,8 +5957,15 @@ function doJsonRequest($w, $url, $actionMode = true) {
 function killUpdate($w) {
 	deleteTheFile($w->data() . '/update_library_in_progress');
 	deleteTheFile($w->data() . '/download_artworks_in_progress');
-	deleteTheFile($w->data() . '/library_new.db');
-	deleteTheFile($w->data() . '/library_old.db');
+
+
+	if (file_exists($w->data() . '/library_old.db')) {
+	    rename($w->data() . '/library_old.db', $w->data() . '/library.db');
+    }
+
+	if (file_exists($w->data() . '/library_new.db')) {
+		deleteTheFile($w->data() . '/library_new.db');
+	}
 
 	exec("kill -9 $(ps -efx | grep \"php\" | egrep \"update_|php -S localhost:15298|ADDTOPLAYLIST|UPDATE_|DOWNLOAD_ARTWORKS\" | grep -v grep | awk '{print $2}')");
